2018 West LA Crash Data
This is a trimmed version of the data. I made it smaller because it is very difficult to map such large data set on a laptop so chris suggested we just do single year. I filtered the data to just be 2018 which still has like 2000 cells. There are still some error in this code that i'm trying to figure out, maybe yall can help.